DAT-344 - Log clicks on search results
When a user clicks on a search result, log the timestamp, search
parameters, sort order, and the index of the selected result in a file.

At present, this information will populate a file called search_logs.csv
in the container's app directory but we will probably want to change
that. This example shows the log file structure:

|time                      |query |sort                              |org|tags      |format|licence|package-id                                    |index|
|--------------------------|------|----------------------------------|---|----------|------|-------|----------------------------------------------|-----|
|2023-05-30 13:10:32.915699|test  |score desc, metadata_modified desc|   |          |      |       |covid-19-online-diary                         |21   |
|2023-05-30 13:12:20.372644|survey|views_recent desc                 |ons|employment|CSV   |ogl-v2 |hours-worked                                  |4    |

Note that only clicks on search results are logged, so selecting an item
from the default dataset page, even if it is filtered for organisation
or sorted by a custom field, does not add a line to the file.