DAT-350 Add API endpoint to download log file
Adds an endpoint at `/api/action/search_logs` that serves the log file
containing user searches and selected datasets. This endpoint is
available only to admin users.

Because all action endpoint responses are wrapped in a standard response
format automatically, the response will not be a csv format file attachment
- it is nested within the `result` key of a json response.