DAT-351 - Download search logs from admin page
Adds a link, in the sidebar of the administrator page, to download the
search logs as a CSV file. Ideally, there would be an additional tab in
the administrator console for this, but this proved complicated so I
went for the simplest possible way to get the functionality in there for
now.