DAT-338 - Make resource link go directly to the file
Where resources are listed on a dataset page, the links currently go to the
resource detail page, and a dropdown on the right allows you to either go to the
resource page or go directly to the resource. This change means that the link
goes directly to the resource, and replaces the dropdown with a single button
that links to the resource detail page.