DAT 323 - Fix data quality boost
The previous implementation incorrectly boosted any results where the data
quality was between 0 and 5 equally, instead of boosting based on the actual
value. This fixes it by adding a boost for each possible value. At some point,
we may want to use function queries to boost based on the data quality value
instead, but luqum does not support this so I have gone with the simplest option
for now.