DAT-371 - Add dataset boost field
Adds a field labelled "boost factor" to datasets, which is editable by admins
but not shown to the public on the dataset page. The purpose of this is so that
admins can specify datasets that should be promoted or demoted within the search
results by entering a positive or negative number, but this change does not
modify the search. I've allowed any number as input but it will be rounded to
4dp; not sure if this is right since I don't know how the field will be
populated but we can change the criteria easily.