DAT-371 Use extra boost field in search
The requirement for this ticket is for admins to be able to insert any
numerical value into the extra field that I've named "dataset boost" to
control how high in the search results a field appears. On the surface,
this sounds like the same issue as adding the data quality field but
proved much more complex because data quality accepts a limited number
of options (numbers 1-5) that can be handled seperately, and we can't do
that for an "any number" field.

Firstly, I had to use function queries to include the number in the
search score instead of just looking for each possible value. This meant
moving away from the `luqum` library that I'd used to generate query
strings, since it doesn't support function queries, and instead just
format the query strings.

Doing this raised a problem for both the existing data quality field,
and the new field I just added: to include a field as a `_val_` in the
query, the field needs to be a numeric datatype in the SOLR schema. CKAN
stores all custom fields added to the schema as strings, only converting
to to and from the specified datatype. Overriding this behaviour would
have taken a lot of work as much of CKAN's behaviour depends on it, so I
instead have added a copy of the two numeric fields to the SOLR schema.
The `update_config` plugin function is called at startup, so I have
added a call to the SOLR endpoint that adds a new field with the correct
datatype as a copy of the string field.

Now, when an admin user enters a "data quality" and "dataset boost"
value for a dataset, they are copied to two fields: a text field used by
the app, and a numeric field used by the search. The quality field
contributes to search scoring in the same way as before, and the boost
field just contributes whatever number you put in to the score, with
equal weighting to the text match. This probably isn't what we want
long-term, but playing around with the search debug endpoint will help
figure out what weighting we wish to assign to the boost field.