DAT-402 - Fix admin user search bug
The call to the search logging endpoint 400'ed when the user was logged in as
admin, because the session cookies were being included in the request but the
csrf token wasn't there. Since this endpoint doesn't need any authentication,
this change omits the credentials from the request and resolves the issue we
were seeing, where the redirection  to the dataset page was blocked. As
you can't use "withCredentials = false" for a same site xrh request
(technically you can, but it does nothing), I have switched to a fetch
so the creds are omitted by default.