DAT-402 - Make search log endpoint anonymous access
Rather than not sending the credentials with the request, which was slow, this
fixes the bug by setting the log selected search result endpoint to ignore the
credentials using a toolkit decorator. The docs suggested that this isn't the
correct use case for this decorator, but it seems to work so I went with it.