DAT-353 Remove unnecessary tabs from user profile page
For this ticket, we needed to remove the button that says "add dataset" and
similarly the "you dont' have any datasets, you can add one" text from the
dataset tab on the profile page.

When I was looking at this, I noticed that we dont' need the dataset tab at all
since users should not be creating and owning any datasets, so I removed it. I
also removed other tabs that we don't need including groups and activity stream
from this page - we don't have any groups and the activity stream is shown
elsewhere on the dashboard. I left the organisation tab for admin users, since
they will need to upload organisation images, but removed it from other users'
profile.