DAT-354 - Remove "add dataset" button from dataset page
For admins, this button appeared above the list on the dataset page but we don't
want to allow the manual addition of datasets so I have removed it. The API
endpoint is still there, so if we need to prevent access in future then we'll
have to block it in nginx on the prod server.