DAT-309-243 Add LGOV scripts and GA
Adds the following scripts:
* Cookie consent manager
* Crisis communication
* Google analytics

It also adds a link to the footer that opens the cookie manager panel.

To avoid sending a bunch of google analytics requests during local development,
these are not added to the page head on localhost. You'll see the link in the
footer, but it won't do anything on localhost.

I've configured the cookie manager with some options that seemed sensible to me
because the default looked a little confusing in my opinion.