DAT-352 - Use readonly instead of disabled for locked dataset fields
Since the previous commit, a change to data quality or boost factor resulted in
the dataset name disappearing. This is because disabled field values are not
sent with the form data, and CKAN overwrites the dataset fields with empty values when
this happens. Therefore I have changed them to be readonly instead of disabled.