Set 'image' icon to 'png'
CKAN does not have an icon for "image" format resources, so the easiest thing is just to use the "png" one for now.
Datapress doesn't expose the original resource filename so it's impossible to know what format the resource actually is without downloading it.
Something like this could work:
https://github.com/ahupp/python-magic
E.g. download the first 2048 bytes and guess the file type