Update author info on dataset page
The default CKAN additional_info page doesn't have individual blocks for
each of the additional info items, so I've had to copy across the
entire template and just change the bits I needed.

The only change to the CKAN template is an extra {% if %} block inside
the author_email row. If there's no author name given then the email
address itself will be displayed in the table cell.