Fix "data_quality: missing_value" error
When using the undelete button on a dataset that has a `data_quality` value of
`None`, we were seeing a missing value error; this was also happening when
adding a dataset or resource despite the schema permitting `None` as a value for
the custom field.

A perplexing issue because using `dataset_update` to edit a dataset with the
exact same values worked just fine - unfortunately I have not yet figured out
why this was happening, but have fixed it by changing the validators for the
field in question.

The problem seems to have been with the `convert_to_extras` converter, which
cannot handle empty fields. As we see in the [docstring for the ignore empty
converter](https://github.com/ckan/ckan/blob/60faa6b5061b3c9dd3ef6971245ef8a5e4d14c16/ckan/lib/navl/validators.py#L235),
CKAN needs the `ignore_empty` or `ignore_missing` validator to appear __before__
the other validators or converters and works by omitting other
validation/conversion steps appearing after it. It's no good having a
validator that can return `None` because the `convert_to_extras` can't handle
this value.

An additional note, because it's not obvious: the `ignore_missing` validator did
not work in this case, and won't work for any form input field, because the
value passed from the form is the empty string, not `None`. It's `ignore_empty`
that we should be using for this and all future optional form fields.