Add additional fields to datasets
Adds the following fields to the dataset edit form and dataset view page:

- project name
- project url
- entry type (dataset or analysis - default is dataset)
- search description

Allows you to filter the search by either project name or entry type.

This does not complete all the criteria for DAT-543 because:

* Harvested datasets don't have a value for entry type as they are not added via
the form. It's not straightforward to make the search filter look for "either x
or null" in the field, so we will have to add the default to the harvester code.

* We don't have the harvest source field required. This will need to be added
seperately as it's a change to the harvester, not the dataset schema.

* I have not added any fields to the resources

* The dataset page UI leaves something to be desired