Override dataset last_modified timestamps
We only want to update a dataset's last modified time if a resource has changed.
Updates to the dataset metadata should not cause the last modified time to be updated.

I've updated the `timestamps.restore_upstream` function (and renamed it to `override`) so that
it'll either use the upstream modified & created times if they exist. Otherwise it'll
get the dataset's resources and find the most recently modified one, and use that
timestamp for the dataset's last_modified timestamp.

Note: A resource's URL is considered part of it's metadata. So updating a
resource's URL will not update the last modified time. I don't think there's
any easy way around this.