Remove groups and showcase tabs
The read_base template is overridden in both the showcase and activity
extensions. The `dataset_type` is added by the base template, and `activity.package_activity`
is added by the activity plugin. The showcase extension also tries to
add its own tab.

Because our extension is loaded first I think it then doesn't also pull
in the changes from the other extensions. But that means we need to
explicitly include the activity tab, because it no longer gets added by
the activity extension.

The CKAN plugin system is still a mystery to me.