Resource page 'additional information' table.
Converts the additional information table on the resource page to use a
description list instead of a table, in the same way as commit #7fe97f6a9b16d3be57bb58b15577b6f70fbc14cb
did for the dataset page. This removes the 'view more' option and shows the
whole table all at once as well as slightly changing the style.

This involved refactoring the template for the dataset page table so a common
snippet is used for both pages.

In addition to the change from table to description list, I have made some edits
to the information shown in the table for a resource:

* Instead of writing "unknown" in place of missing values, omit the row
entirely. This makes it consistent with the dataset page.

* Remove extra fields from the table: we previously had all sorts of extra
fields in the expandable part of the table that didn't seem useful, such as the
boolean `hasViews` and the index of the resource in the resoucre list.
I have changed it so only the file size is retained from this section.