Add skip link to all pages with sidebar
This modifies the base page template so that all sidebars across the site begin
with a "skip to main content" link that takes you to the main page content. The
purpose is to make it easier to navigate using a screen reader, so you don't
need to go through the entire sidebar content before getting to the main page content.