Add search page specific main content skip link
The previous commit made it possible for screenreader users to skip
to the main content from the top of the sidebar on all pages, using a link that
says "skip to main content". This change adds a more specific link for the
search page that says "skip to search results". It does exactly the same thing,
but is a little more informative than the generic version.