Show more of facet filter option text
The text in the search filter options lists (such as organisation
names) was being truncated to 22 characters, full words only. This meant that we
weren't seeing enough text to know what the filters mean, for example with
several organisations shown as "London Borough of..." or "Office for...".

I have now changed it so that:

* We show up to 30 characters (just shy of the maximum number we could show on a
single line here)

* Partial words are shown when the 30-character split happens in the middle of
the word.