Update footer structure
Results of accessibility testing required that the footer header should
be a h2 and the footer links should be in one list rather than 2.

Only minor changes to the template and css were needed to fix this.

I've added an 'h4' class to the h2 so that it looks exactly the same
as before.