DAT-427 - Change sort order selector to links
When tab-selecting the "order by" selector on the search page, the page is
automatically reloaded when you select an option. We want users to be able to
see the options without applying them so I have changed the selector to a list
of links, with the currently active sort order highlighted in pink.