DAT-551 Add custom fields to resource
Adds some fields to the resource schema, as specified in the ticket. None of the
fields are mandatory but can be added by a user with edit access, either in the
website or via API. They are displayed in the 'additional information' table on
the resource detail pages but do not affect the parent dataset in any way.