Move boosts from q parameter to bf parameter
This is functionaly identical in terms of ranking to the existing
behaviour. In particular it has been confirmed that the bf (boost
function) parameter is additive like the previous _val_ syntax.

The purpose of this change is a small refactoring to retain support
for the existing boosting mechanism, whilst also returning to SOLR
default behaviour around the interpretation of the `q` parameter (used
for text search matching). This work will be carried out in a future
commit.

Explaining results in SOLR with debug mode on returns the same
calculation for the same set of results before and after the query.

An example explain output used for testing is shown below:

16.61907 = sum of:
  0.6190695 = weight(text:popul in 12) [SchemaSimilarity], result of:
    0.6190695 = score(freq=19.0), computed as boost * idf * tf from:
      1.5974071 = idf, computed as log(1 + (N - n + 0.5) / (n + 0.5)) from:
        242 = n, number of documents containing term
        1197 = N, total number of documents with field
      0.38754648 = tf, computed as freq / (freq + k1 * (1 - b + b * dl / avgdl)) from:
        19.0 = freq, occurrences of term within document
        1.2 = k1, term saturation parameter
        0.75 = b, length normalization parameter
        36888.0 = dl, length of field (approximate)
        1116.8262 = avgdl, average length of field
  0.5 = FunctionQuery(int(copy_data_quality)), product of:
    5.0 = int(copy_data_quality)=5
    0.1 = boost
  15.5 = FunctionQuery(double(copy_dataset_boost)), product of:
    15.5 = double(copy_dataset_boost)=15.5
    1.0 = boost

In this example we can see a query of "population" (stemmed) where the
item had a dataset_quality of 5 and a dataset boost of 15.5.

- Data quality is scaled 1-5 and is multiplied by 0.1
- Dataset boost is an unscaled float and is multiplied by 1 (used as
- is)

Hence in the above example 15.5 and 0.5 are added as additional boosts
to the score.

I suspect it would be nicer if these boosts were multiplicative
factors of the relevance match, as that way it would allow the boost
to enhance the relevance of the match scores rather than just always
be added to the datasets score regardless of how much it matches.

A future piece of work may look at this a good article on the subject
is here:

https://nolanlawson.com/2012/06/02/comparing-boost-methods-in-solr/