Make sorting case insensitive and improve non alphanumeric sorting
Prior to this commit as part of the work on this PR alphabetical
sorting used CKANs title_string field which appeared to be meant for
this purpose:

https://github.com/ckan/ckan/blob/8ad65497cba82244317ffe6d90fe0d43070bdc97/ckan/config/solr/schema.xml#L163

Unfortunately it was noticed this would sort case sensitively, so
upper case letters would sort before lower case, and non alphanumeric
characteres would be grouped also.

There are some poor quality datasets with titles all lower cased, and
one dataset called "'Climate Just' data" note the single quotes in its
name, that would cause this to be sorted first.

This commit fixes this by introducing a new field_type called
dfl_title_sort which defines a customer analyzer that strips non
alphanumeric characters and downcases the whole string for sorting
comparison.  This leads to more consistent sorting behaviour in the
above edge cases.

Datasets that have titles consisting of only non-alphanumeric
characters will be indexed as the empty string and sort together.