Fix DAT-843 email verification message shown when validation fails
Following this commit the above will no longer occur, and the
validation failure will force the user to fix the validation errors
before we state that an email has been sent.

Also fixes DAT-835 which is a dupe of DAT-843.

Also fixes DAT-837 which is a symptom of accounts not being created
due to DAT-843/835.